+++ packages/components/index.ts
@@ -40,11 +40,8 @@ export * from "./checkbox-input";
 export * from "./checkbox-input-number";
 
 export * from "./checkbox-group-input";
-
 export * from "./checkbox-group";
-
 export * from "./checkbox-date-picker";
-
 export * from "./checkbox-date-range-picker";
-
 export * from "./checkbox-group-input-number";
+export * from "./uploader-uppy";