+++ packages/components/uploader-uppy/index.ts
@@ -1,0 +1,9 @@
+import { withInstall } from "@viaz/utils";
+
+import UploaderUppy from "./src/index.vue";
+
+
+export const VzUploaderUppy = withInstall(UploaderUppy);
+
+
+export default VzUploaderUppy;