+++ packages/types/components/index.ts
@@ -57,3 +57,5 @@ export * from "./checkbox-group";
 export * from "./checkbox-date-picker";
 export * from "./checkbox-date-range-picker";
 export * from "./checkbox-group-input-number";
+
+export * from "./uploader-uppy";