+++ packages/types/components/uploader-uppy.ts
@@ -1,0 +1,5 @@
+export interface VzUploaderUppyProps {
+  /** 上传服务接口地址 */
+  url?: string;
+
+}