+++ packages/components/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @viaz/components
 
+## 0.1.13
+
+### Patch Changes
+
+- fixed bug
+- Updated dependencies
+  - @viaz/constants@0.1.13
+  - @viaz/enums@0.1.13
+  - @viaz/hooks@0.1.13
+  - @viaz/types@0.1.13
+  - @viaz/utils@0.1.13
+
+## 0.1.12
+
+### Patch Changes
+
+- update ARangePicker presets
+- Updated dependencies
+  - @viaz/constants@0.1.12
+  - @viaz/enums@0.1.12
+  - @viaz/hooks@0.1.12
+  - @viaz/types@0.1.12
+  - @viaz/utils@0.1.12
+
+## 0.1.11
+
+### Patch Changes
+
+- add inited event
+- Updated dependencies
+  - @viaz/constants@0.1.11
+  - @viaz/enums@0.1.11
+  - @viaz/hooks@0.1.11
+  - @viaz/types@0.1.11
+  - @viaz/utils@0.1.11
+
 ## 0.1.10
 
 ### Patch Changes